FE: Bake dashboard into ws board page
